Backend/requirements.txt
fastapi==0.115.8
matplotlib==3.10.1
numpy==1.26.4
langchain==0.3.26
langchain-core==0.3.69
langchain-openai==0.3.28
openai-whisper==20240930
openpyxl==3.1.5
pandas==2.2.3
protobuf==4.25.3
pydantic==2.10.3
pyodbc==5.2.0
python-dotenv==1.0.1
RapidFuzz==3.12.2
scikit-learn==1.6.1
sentence-transformers==3.4.1
transformers==4.49.0
wheel==0.45.1
python-dotenv==1.0.1
xhtml2pdf==0.2.17
python-multipart==0.0.20
uvicorn==0.29.0
huggingface_hub[hf_xet]